pyproject.toml
[tool.poetry]
name = "flake8_ado"
version = "0.1.0"
description = "Flake8 plugin to check comments against AzureDevOps ticket references."
readme = "README.md"
homepage = "https://github.com/DanielKusyDev/flake8-ado"
repository = "https://github.com/DanielKusyDev/flake8-ado"
keywords = [
    "flake8",
    "aod",
    "azure devops",
    "lint"
]
authors = ["DanielKusyDev <daniel.kusy97@gmail.com>"]
license = "MIT"

classifiers = [
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Topic :: Software Development :: Quality Assurance",
]

[tool.poetry.dependencies]
python = "^3.7"
azure-devops = "^6.0.0-beta.4"
importlib-metadata = "4"
pytest-mock = "^3.8.2"
tox = "^3.26.0"

[tool.poetry.group.dev.dependencies]
isort = "^5.10.1"
black = "^22.8.0"
flake8-black = "^0.3.3"
pytest = "^7.1.2"

[tool.black]
line-length = 120
target-version = ['py37']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.mypy_cache
    | \.pytest_cache
    | \.tox
    | dist
  )/
)
'''

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120

[tool.poetry.plugins]
[tool.poetry.plugins."flake8.extension"]
ADO = "flake8_ado:Plugin"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
